+++ ImTK/UI/Element/FieldDrawer/DrawerLayoutMode.cs
@@ -1,0 +1,8 @@
+namespace ImTK.UI
+{
+    public enum DrawerLayoutMode
+    {
+        Inline,
+        Expand
+    }
+}