+++ ImTK/UI/Event/IValueChangedEvent.cs
@@ -1,0 +1,10 @@
+namespace ImTK.UI
+{
+    public interface IValueChangedEvent
+    {
+        object previousValueObj { get; }
+        object newValueObj { get; }
+        bool isInternalChange { get; }
+        bool bubbles { get; }
+    }
+}